blend probabilities

+++ src/encode/factor.rs
@@ -1,3 +1,3 @@
-pub fn shrinkage_factor(num_case: i32, param: f32) -> f32 {
-    num_case as f32 / (num_case as f32 + param)
+pub fn shrinkage_factor(num_case: i64, param: f32) -> f32 {
+    (num_case as f64 / (num_case as f64 + param as f64)) as f32
 }
